// generated/src/app/api-models/pcpaa-rules-hdr.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PcpaaRulesHdr {
  ruleId: string;
  ruleType: number;
  ruleDescr: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}